fix(geo): address review feedback on own-brand row and Logo Link helper
Move the Logo Link URL builder into @notra/utils so geo-core and ui share
one implementation, normalize alias and domain forms before excluding the
own website from synonyms, and drop manual memoization in the competitors
table now that React Compiler manages it.
